move SliderUiState to model package
