+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/designSystem/component/Slider.kt
@@ -26,11 +26,8 @@ import androidx.compose.ui.unit.dp
 import com.sanaa.tudee_assistant.R
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.Theme
 import com.sanaa.tudee_assistant.presentation.designSystem.theme.TudeeTheme
-import com.sanaa.tudee_assistant.presentation.model.TaskUiState
-import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
+import com.sanaa.tudee_assistant.presentation.model.SliderUiState
 import com.sanaa.tudee_assistant.presentation.model.TudeeUiStatus
-import com.sanaa.tudee_assistant.presentation.screen.home.SliderUiState
-import com.sanaa.tudee_assistant.presentation.utils.DataProvider
 
 @Composable
 fun Slider(
